+++ src/main/java/com/epam/javacourses2016/task13/AbstractGraphCreator.java
@@ -10,8 +10,12 @@ public abstract class AbstractGraphCreator {
      * @param numberNodes Количество вершин в графе.
      * @return Граф указанной конфигурации.
      */
-    public static AbstractGraph createGraph(int numberNodes){
-        return null;
+    public static AbstractGraph createGraph(int numberNodes) {
+
+        if (numberNodes <= 0) {
+            return null;
+        }
+        return new MyGraph(numberNodes);
     }
 
     /**
@@ -20,9 +24,11 @@ public abstract class AbstractGraphCreator {
      * Нумерация вершин начинается с 0.
      * Допустимы операции добавления, удаления и проверки существования ребер.
      */
-    abstract class AbstractGraph {
+    static abstract class AbstractGraph {
 
-        /** Количество вершин */
+        /**
+         * Количество вершин
+         */
         protected final int NUMBER_NODES;
 
         public AbstractGraph(int numberNodes) {
@@ -31,21 +37,24 @@ public abstract class AbstractGraphCreator {
 
         /**
          * Добавление ребра в граф.
-         * @param first Первая связываемая вершина.
+         *
+         * @param first  Первая связываемая вершина.
          * @param second Вторая связываемая вершина.
          */
         public abstract void addEdge(int first, int second);
 
         /**
          * Удаление ребра из графа.
-         * @param first Первая освобождаемая от связи вершина.
+         *
+         * @param first  Первая освобождаемая от связи вершина.
          * @param second Вторая освобождаемая от связи вершина.
          */
         public abstract void removeEdge(int first, int second);
 
         /**
          * Проверка наличия ребра.
-         * @param first Первая вершина.
+         *
+         * @param first  Первая вершина.
          * @param second Вторая вершина.
          */
         public abstract boolean isExistEdge(int first, int second);
