+++ src/main/java/com/epam/javacourses2016/task14/MyCollection.java
@@ -3,10 +3,10 @@ package com.epam.javacourses2016.task14;
 import java.util.*;
 
 public class MyCollection<T extends Number> extends AbstractCollection<T> implements AbstractCollectionCreator.NumberCollection<T> {
-    List<T> elementData;
+    private List<T> elementData;
 
     public MyCollection() {
-        elementData = new ArrayList<T>();
+        elementData = new ArrayList<>();
     }
 
     @Override
@@ -40,5 +40,10 @@ public class MyCollection<T extends Number> extends AbstractCollection<T> implem
     @Override
     public boolean addAll(Collection<? extends T> c) {
         return elementData.addAll(c);
+    }
+
+    @Override
+    public boolean add(T t) {
+        return elementData.add(t);
     }
 }
