+++ src/main/java/com/epam/javacourses2016/task5/SolverTask5.java
@@ -1,6 +1,7 @@
 package com.epam.javacourses2016.task5;
 
 import java.math.BigDecimal;
+import java.math.RoundingMode;
 import java.util.List;
 
 /**
@@ -18,6 +19,10 @@ public class SolverTask5 {
      * @return Вычисленное по исходным данным сопротивление.
      */
     double calcResistance(List<Measurement> measurements) {
+        if (measurements == null || measurements.size() == 0) {
+            return 0;
+        }
+
         double xSum = 0;
         double ySum = 0;
         double xSqrSum = 0;
@@ -32,9 +37,12 @@ public class SolverTask5 {
 
         double measureCount = measurements.size();
 
+        if (xSum == 0) {
+            return 0;
+        }
+
         BigDecimal aBig;
         aBig = new BigDecimal((measureCount * xySum - (xSum * ySum)) / (measureCount * xSqrSum - xSum * xSum));
-
-        return aBig.doubleValue();
+        return aBig.setScale(3, RoundingMode.HALF_UP).doubleValue();
     }
 }
