+++ src/test/java/com/epam/javacourses2016/Task17/SolverTask17Test.java
@@ -13,7 +13,7 @@ import java.util.Set;
 
 public class SolverTask17Test {
 
-    @Test(enabled = false, dataProvider = "segments")
+    @Test(enabled = true, dataProvider = "segments")
     public void testSegmentsIntersection(ArrayList<double[][]> segments, double[][] points) {
         SolverTask17 solver = new SolverTask17();
