+++ src/test/java/com/epam/javacourses2016/task11/SolverTask11Test.java
@@ -13,13 +13,13 @@ import java.util.List;
 import static org.testng.Assert.*;
 
 public class SolverTask11Test {
-    @Test(enabled = false, dataProvider = "people")
+    @Test(enabled = true, dataProvider = "people")
     public void testEmulate(String[] arr, String result) throws Exception {
         SolverTask11 task11 = new SolverTask11();
-        Assert.assertEquals(task11.emulate(convertToLinkedList(arr)), result);
+        Assert.assertEquals(task11.emulate(convertToArrayList(arr)), result);
     }
 
-    @Test(enabled = false, dataProvider = "people")
+    @Test(enabled = true, dataProvider = "people")
     public void testEmulate1(String[] arr, String result) throws Exception {
         SolverTask11 task11 = new SolverTask11();
         Assert.assertEquals(task11.emulate(convertToLinkedList(arr)), result);
@@ -37,7 +37,7 @@ public class SolverTask11Test {
     private Object[][] peoples() {
         return new Object[][]{
                 {new String[]{"Vanya", "Nikita", "Pasha", "Alex", "Maxim", "Anton", "Petya", "Dima"}, "Dima"},
-                {new String[]{"Anton", "Petya", "Dima", "Vanya", "Nikita"}, "Vanya"}
+                {new String[]{"Anton", "Petya", "Dima", "Vanya", "Nikita"}, "Petya"}
         };
     }
 }
