+++ src/test/java/com/epam/javacourses2016/task13/AbstractGraphCreatorTest.java
@@ -7,7 +7,7 @@ import org.testng.annotations.Test;
 import static org.testng.Assert.*;
 
 public class AbstractGraphCreatorTest {
-    @Test(enabled = false, dataProvider = "nodes")
+    @Test(enabled = true, dataProvider = "nodes")
     public void testCreateGraph(int numberNodesResult) throws Exception {
         AbstractGraphCreator.AbstractGraph graph = AbstractGraphCreator.createGraph(numberNodesResult);
         if(numberNodesResult<=0){
