+++ src/test/java/com/epam/javacourses2016/task14/NumberCollectionTest.java
@@ -11,7 +11,7 @@ import static org.testng.Assert.*;
  */
 public class NumberCollectionTest {
 
-    @Test(enabled = false, dataProvider = "numbers")
+    @Test(enabled = true, dataProvider = "numbers")
     public void testNearest(Number[] numbers, Number value, Number result) throws Exception {
         AbstractCollectionCreator.NumberCollection<Number> collection = AbstractCollectionCreator.createCollection(Number.class);
         for (Number n: numbers) collection.add(n);
