+++ src/test/java/com/epam/javacourses2016/task18/SolverTask18Test.java
@@ -9,7 +9,7 @@ import org.testng.annotations.Test;
  */
 public class SolverTask18Test {
 
-    @Test(enabled = false, dataProvider = "matrix")
+    @Test(enabled = true, dataProvider = "matrix")
     public void testGetMaxSubMatrix(int[][] data, int[][] result) throws Exception {
         SolverTask18 task18 = new SolverTask18();
         SolverTask18.RectangularIntegerMatrix matrix = task18.getMaxSubMatrix(convertToMatrix(data));
